+++ src/main/java/until/the/eternity/dcs/domain/announcement/infrastructure/AnnouncementRepository.java
@@ -2,9 +2,11 @@ package until.the.eternity.dcs.domain.announcement.infrastructure;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 import until.the.eternity.dcs.domain.announcement.entity.Announcement;
 
-public interface JpaAnnouncementRepository extends JpaRepository<Announcement, Long> {
+@Repository
+public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
 
     Boolean existsByPostId(Long postId);
