+++ src/test/java/until/the/eternity/dcs/domain/announcement/application/AnnouncementServiceTest.java
@@ -17,8 +17,8 @@ import until.the.eternity.dcs.domain.announcement.dto.request.AnnouncementCreate
 import until.the.eternity.dcs.domain.announcement.dto.response.AnnouncementPageResponseItem;
 import until.the.eternity.dcs.domain.announcement.dto.response.AnnouncementPersistResponse;
 import until.the.eternity.dcs.domain.announcement.entity.Announcement;
-import until.the.eternity.dcs.domain.announcement.entity.AnnouncementRepository;
 import until.the.eternity.dcs.domain.announcement.exception.AnnouncementDuplicateException;
+import until.the.eternity.dcs.domain.announcement.infrastructure.AnnouncementRepository;
 import until.the.eternity.dcs.domain.board.entity.Board;
 import until.the.eternity.dcs.domain.post.entity.Post;
 import until.the.eternity.dcs.domain.post.infrastructure.PostMetaRepository;
@@ -122,7 +122,8 @@ class AnnouncementServiceTest {
     @DisplayName("getAnnouncementByBoardId는 BoardId로 해당 게시판과 전체 공지글을 조회한다.")
     void getAnnouncementByBoardId_Success() {
         // given
-        when(announcementRepository.findByBoardIdAndGlobal(id)).thenReturn(List.of(announcement));
+        when(announcementRepository.findByBoardIdOrIsGlobalTrue(id))
+                .thenReturn(List.of(announcement));
 
         // when
         List<AnnouncementPageResponseItem> response =
